Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
     const float pi=3.14;
-    double radius;
-    scanf("%lf",&radius);
+    float radius;
+    scanf("%f",&radius);
     printf("Area: %.2f",pi*radius*radius);
     return 0;
 }
